+++ example/markers/app.js
@@ -6,8 +6,7 @@ var wavesurfer; // eslint-disable-line no-var
 // Init & load audio file
 document.addEventListener('DOMContentLoaded', function() {
     // Init
-    var button = document.createElement("button");
-    button.innerHTML = "click";
+
     wavesurfer = WaveSurfer.create({
         container: document.querySelector('#waveform'),
         waveColor: '#A8DBA8',
@@ -26,13 +25,7 @@ document.addEventListener('DOMContentLoaded', function() {
                         label: "V1",
                         color: '#ff990a'
                     },
-                    {
-                        time: 10,
-                        label: "V2",
-                        color: '#00ffcc',
-                        position: 'top',
-                        markerElement: button
-                    },
+
                     {
                         time: 24,
                         label: "END",
@@ -43,6 +36,16 @@ document.addEventListener('DOMContentLoaded', function() {
             })
         ]
     });
+
+    var img = new Image(40, 40);
+    img.src = "./settings_icon.png";
+    img.onload = () => {
+        wavesurfer.markers.add({
+            time: 12,
+            position: "bottom",
+            markerElement: img
+        });
+    };
 
     wavesurfer.on('error', function(e) {
         console.warn(e);